change Dimness window collection behavior
canJoinAllSpaces is needed for it to work across all Spaces on all
monitors. However that will cause black background (of Dimness Window)
when switching to another Space and the Desktop overview is also black.

Change to this collection behavior fixes it but then it won't work for
Spaces other than the current shown one.